agent/requirements.txt
# LangGraph and LangChain dependencies
langgraph>=0.2.74
langchain>=0.3.0  
langchain-core>=0.3.0
langchain-openai>=0.2.0
langsmith>=0.1.0

# OpenAI integration
openai>=1.0.0

# Azure authentication
azure-identity>=1.15.0
azure-core>=1.29.0

# FastAPI and web framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# HTTP client
httpx>=0.27.0
aiohttp>=3.9.0

# Database and caching
asyncpg>=0.29.0
redis>=5.0.0

# Environment and utilities
python-dotenv>=1.0.0
